front

+++ clients/src/main/java/com/EricFeng/clients/fraud/FraudClient.java
@@ -1,8 +1,10 @@
 package com.EricFeng.clients.fraud;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+
 
 @FeignClient(
         name = "fraud",
@@ -10,7 +12,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface FraudClient {
 
         @GetMapping(path = "api/v1/fraudCheck/{customerId}")
-        CheckFraudResponse CheckFraudHistory(@PathVariable("customerId") Integer customerId);
+
+        CheckFraudResponse CheckFraudHistory(@PathVariable("customerId") Long customerId);
 }
